Fix up warnings and make general improvements (#19)
* Fix up warnings
* Fix linux issues with time
* Fix some memory leaks
* Change output directory for generated messages
* Setup message generation to properly handle namespacing
* Improve help messages
* Move pubsub_cpp includes

+++ include/pubsub/System.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include <inttypes.h>
 #ifdef __cplusplus
 extern "C"
 {
@@ -7,7 +7,7 @@ extern "C"
 
 void ps_sleep(unsigned int time_ms);
 
-unsigned long long GetTimeMs();
+uint64_t GetTimeMs();
 
 #ifdef __cplusplus
 }
